8.25 fourth Revision qly

+++ src/main/java/com/dicadut/soms/entity/Wy.java
@@ -6,18 +6,17 @@ import lombok.Data;
 import java.time.LocalDateTime;
 
 /**
- * @author Radium
- * @version 1.0
- * @date 2021-04-13 17:25:56
+ * @author Yang
+ * @Description TODO
+ * @create 2021-08-25 16:59
  */
 @Data
 public class Wy {
     @TableId
     private Integer id;
-    private Long serial;
+    private Integer serial;
     private LocalDateTime getTime;
     private String loraRawData;
     private String wyRawData;
     private Double value;
-
 }
